+++ src/cli/commands/sdi-datalakes.ts
@@ -8,6 +8,7 @@ import {
     getColor,
     getSdk,
     homeDirLog,
+    printObjectInfo,
     proxyLog,
     serviceCredentialLog,
     verboseLog,
@@ -19,7 +20,7 @@ let color = getColor("magenta");
 
 export default (program: CommanderStatic) => {
     program
-        .command("sdi-datalakes")
+        .command("sdi-data-lakes")
         .alias("sdl")
         .option(
             "-m, --mode [list|create|update|template|info|delete]",
@@ -76,13 +77,13 @@ export default (program: CommanderStatic) => {
         })
         .on("--help", () => {
             log("\n  Examples:\n");
-            log(`    mc sdi-datalakes --mode list \t\t list all sdi datalakes`);
-            log(`    mc sdi-datalakes --mode template \t\t create template file`);
-            log(`    mc sdi-datalakes --mode create --datalake <datalakefile> \t\t create sdi data lake`);
-            log(`    mc sdi-datalakes --mode update --datalake <datalakefile> --datalakeid <datalakeid> \
+            log(`    mc sdi-data-lakes --mode list \t\t list all sdi datalakes`);
+            log(`    mc sdi-data-lakes --mode template \t\t create template file`);
+            log(`    mc sdi-data-lakes --mode create --datalake <datalakefile> \t\t create sdi data lake`);
+            log(`    mc sdi-data-lakes --mode update --datalake <datalakefile> --datalakeid <datalakeid> \
                                                                               \t\t update sdi data lake`);
-            log(`    mc sdi-datalakes --mode info --datalakeid <datalakeid>   \t\t get sdi data lake info`);
-            log(`    mc sdi-datalakes --mode delete --datalakeid <datalakeid> \t\t delete sdi data lake`);
+            log(`    mc sdi-data-lakes --mode info --datalakeid <datalakeid>   \t\t get sdi data lake info`);
+            log(`    mc sdi-data-lakes --mode delete --datalakeid <datalakeid> \t\t delete sdi data lake`);
 
             serviceCredentialLog();
         });
@@ -92,28 +93,28 @@ function checkRequiredParamaters(options: any) {
     options.mode === "create" &&
         !options.datalake &&
         errorLog(
-            "you have to provide a datalake template file to create a sdi datalake (see mc sdi-datalakes --help for more details)",
+            "you have to provide a datalake template file to create a sdi datalake (see mc sdi-data-lakes --help for more details)",
             true
         );
 
     options.mode === "update" &&
         !options.datalakeid &&
         errorLog(
-            "you have to provide the datalakeid of the datalake you want to update (see mc sdi-datalakes --help for more details)",
+            "you have to provide the datalakeid of the datalake you want to update (see mc sdi-data-lakes --help for more details)",
             true
         );
 
     options.mode === "info" &&
         !options.datalakeid &&
         errorLog(
-            "you have to provide the datalakeid to get infos about the sdi data lake (see mc sdi-datalakes --help for more details)",
+            "you have to provide the datalakeid to get infos about the sdi data lake (see mc sdi-data-lakes --help for more details)",
             true
         );
 
     options.mode === "delete" &&
         !options.datalakeid &&
         errorLog(
-            "you have to provide the datalakeid to delete the sdi data lake (see mc sdi-datalakes --help for more details)",
+            "you have to provide the datalakeid to delete the sdi data lake (see mc sdi-data-lakes --help for more details)",
             true
         );
 }
@@ -142,15 +143,7 @@ function printDataLakeInfos(
     dataLake: SemanticDataInterconnectModels.DataLakeResponse | SemanticDataInterconnectModels.CreateDataLakeResponse,
     options: any
 ) {
-    console.log("DataLake");
-    console.log(`Id: ${color(dataLake.id)}`);
-    console.log(`Name: ${color(dataLake.name)}`);
-    console.log(`Type: ${dataLake.type}`);
-    console.log(`Created Date: ${dataLake.createdDate}`);
-    console.log(`Updated Date: ${dataLake.updatedDate}`);
-    console.log(`Base Path: ${color(dataLake.basePath)}`);
-
-    verboseLog(JSON.stringify(dataLake, null, 2), options.verbose);
+    printObjectInfo("Data Lake:", dataLake, options, ["id", "name", "basePath"], color);
 }
 
 function createTemplate(options: any) {
@@ -175,7 +168,7 @@ function writeToFile(options: any, dataLake: any) {
     console.log(
         `The data was written into ${color(
             fileName
-        )} run \n\n\tmc sdi-datalakes --mode create --datalake ${fileName} \n\nto create the sdi data lake`
+        )} run \n\n\tmc sdi-data-lakes --mode create --datalake ${fileName} \n\nto create the sdi data lake`
     );
 }
 async function createDataLake(options: any, sdk: MindSphereSdk) {