+++ CHANGELOG.md
@@ -13,10 +13,13 @@
 - CLI: new `mc mobile-apps` command [#252]
 - CLI: new `mc notifications` command for MindSphere notifications [#252]
 - CLI: new `mc signal-calculation` command for MindSphere Signal Calculation Service [#258]
+- CLI: new `mc sdi-data-lakes` command for SDI [#265]
+- CLI: new `mc sdi-data-registries` command for SDI [#265]
 - CLI: fixed bug in `mc aggregates` command with multiple lines of same aggregate
 - SDK: signal calculation client [#258]
 - SDK: Job Manager Client [#255]
 - SDK: Data Exchange Client [#256]
+- SDK: AnomalyDetection Client [#254]
 - SDK: NotficationV4 Client [#252]
 - SDK: IotTimeSeriesAggragetesV4 Client [#250]
 - Bumped all dependencies